Initial commit
The scope of this project is to build an Android build for the Oneplus 5
phone (Cheeseburger) without changing anything on AOSP. We are 'allowed'
to only add repos, but we cannot change anything on AOSP.

We'll keep track of the latest AOSP build tested with this manifest. The
instructions will be as:
$ repo init -u https://android.googlesource.com/platform/manifest -b android-11.0.0_r3
$ cd .repo
$ git clone https://github.com/robertoglxda/local_manifests.git local_manifests
$ cd ..
$ repo sync -c --no-clone-bundle --no-tags

We aim to be able to change the AOSP tag without changing anything else.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>